Fix nodeConfigDefaults on cluster create.
The expansion of
`node_pool_defaults.node_config_defaults.insecure_kubelet_readonly_port_enabled`
should have a ForceSendFields to allow `FALSE` values to be propagated.

Without this "FALSE" values are not included in the request.

This is not an issue in nodePoolAutoConfig, or during updates.
ForceSendFields is set in those cases.

An alternative fix is to use expandNodeKubeletConfig function here as
well. The server will reject any values other than the readonly port
field.