Update docs on running single tests
The examples refer to running a single test, but the `-run` command
example could potentially run lots of tests unexpectedly.

Use a `-run=` example with the correct syntax for passing a literal
dollar sign to the regex -- while the docs link to the golang docs,
because of shell quoting and because of it getting passed via make, it
seems to be necessary to use double dollar signs.

Fixes hashicorp/terraform-provider-google#19934