Fix Firestore field deletion for wildcard fields
You can update index configs for wildcard fields, but you can't update TTL configs.
When we "delete" SFIs, we actually issue a PATCH with empty index and TTL configs.
We put both indexConfig and ttlConfig in the update mask for this PATCH accordingly.

For wildcard fields, simply exclude ttlConfig from the update mask when issuing this PATCH to avoid the "unsupported" error.