Update cluster versions in TestAccContainerNodePool_concurrent
At time of this writing, the cluster default version is 1.31.5,
so the API will not update such as cluster to 1.29.4.

This change does not "fix" the need to update these static versions
from time to time. It does leave breadcrumbs for the next traveler
to zero in on the answer.