compute: Documented empty string for `load_balancing_scheme`
While it's not listed in the possible enum values, empty string is also
used in certain situations as a valid value for a
compute_forwarding_rule's `load_balancing_scheme`. Since the valid
values are auto-generated, add a note about the empty string as well.

This can be confusing, since using `null` or unsetting the value will
result in the default (`"EXTERNAL"`).

Fixes https://github.com/hashicorp/terraform-provider-google/issues/22098