feat(LbTrafficExtension): Add LBTrafficExtension metadata support on extension level
The GCP LbTrafficExtension supports to give an object of metadata on setup.
This behaviour shall be added with this change.

You can setup an loadbalancer traffic extension with a `traffic.yml` like this:
```yaml
name: queueit-traffic-extension
forwardingRules:
  - https://www.googleapis.com/compute/v1/projects/[PROJECT]/regions/global/forwardingRules/[GCP_FORWARD_RULE]
loadBalancingScheme: EXTERNAL_MANAGED
extensionChains:
  - name: "chain1"
    matchCondition:
      celExpression: 'request.host = 'example.com'
    extensions:
      - name: 'cloud-connector'
        authority: [AUTHORITY]
        service: https://www.googleapis.com/compute/v1/projects/[PROJECT]/regions/global/backendServices/[BACKEND_SERVICE]
        failOpen: true
        timeout: 1s
        supportedEvents:
          - REQUEST_HEADERS
          - REQUEST_BODY
          - RESPONSE_HEADERS
        metadata:
          "customerId": "test"
          "key": "value"
```

Which will be added to metadata in the traffic service extension. The metadata object need to be a JSON object to be valid. Afterwards the command:

```bash
$ gcloud service-extensions lb-traffic-extensions import test-traffic-extension --source=traffic.yaml --location=global
```

will successfully create an extension including the metadata.