Remove datastore in update test
Post-test destory failed with `Error running post-test destroy, there may be dangling resource. Error when reading or editing DataStore: googleapi: Error 400: DataStore ... currently exists in list at index 0 of Engine`.
This is because a discovery engine was created and references the
datastore. We must destory the engine first. However, the engine
ID is stored in `Agent > GenAppBuilderSettings` field, which is
currently missing on Terraform.