Add Network Firewall Policy "Policy Type" field
The policy type field is a new field, set on policy creation and is then
immutable.
Used to define which resources the policy can be associated with
(depending on the VPC type), and imposes limitations and validations on
the rules in the policy.

The default policy type is VPC_POLICY. RDMA_ROCE_POLICY can be used with
RoCE networks.