resource_composer_environment.go.tmpl - handle "empty changes" to RecoveryConfig
Empty ScheduledSnapshotsConfig and config with scheduled snapshots explicitly disabled represent in fact the same configuration. If applying a sub-change (complex environment updates are split into smaller ones - handled directly by the Composer API) fails specifically because it does not bring any actual modification, this error should be silently ignored. This is not feasible to replicate the logic present in Composer API that decides if a change should actually trigger an update operation.