feat: remove `deployedModelIds` field and avoid setting its value  in `post_create`
Remove the `deployedModelIds` field as it no longer needs to be set as a resource field during `post_create`.
Though deployed model Ids are required to undeploy models from endpoint during resource deletion, they can be read with an API call during the delete step and do not need to be set as a resource field during `post_create`.