feat: update `custom_delete` implementation to disallow > 1 model deployed to endpoint.
Since resource is `immutable`, endpoint should always have 1 deployed model and it is not necessary to account for the endpoint having more than one model manually deployed to it outside of Terraform.
Hence, `deployedModelId` of the 1 deployed model can be set as a Terraform field during `post_create` and retrieved later during `delete` to undeploy the model.
`deployedModelDisplayName` field is also added for more helpful debug logging.