Made gkehub feature membership use same custom endpoint as other resources
This resource previously used a special custom endpoint that wasn't shared, but we accidentally broke it in https://github.com/GoogleCloudPlatform/magic-modules/pull/12287. Prior to that, the separate endpoint had already been a source of confusion, so we're standardizing on a shared endpoint per GKEHub service folder to resolve this issue.