Improve root_password and password_validation_policy updates in sql instance
In cases where both the password_validation_policy and the root_password are
being changed - it is important to first update the password_validation_policy
and only then update the password.

Consider the case where mistakenly we set a too strict policy and now the
password updates are failing - we now want to loosen the policy, but we always
fail on updates since the password change always happens first.
This requires now doing a "two-step" terraform apply instead of single one.