resource_test change: do not rely on Getwd
Depending on how the package is imported (e.g. a future with Bazel), `Getwd()` may be many directories above the root directory for MM, which makes this test fail.

Using a heuristic related to the test file however works in both cases. This matches the technique in validate_third_party_test