Make GKE cluster enable_l4_ilb_subsetting a computed field
This change modifies the enable_l4_ilb_subsetting field in the google_container_cluster resource to be a "computed" field. This means that if the value is not explicitly set in the Terraform configuration, the provider will accept the value determined by the GKE control plane instead of defaulting to false. The Terraform provider will not create a diff when the field is not explicitly set in the user's configuration, thus avoiding potential errors and unexpected behavior.