Add data files as embedsrcs for the main MM binary.
* Create two BUILD files which include data files under `third_party` and `templates`
* Add them to `embedsrcs` during the copybara import.

This is so the main binary has all the files available for future inclusion via Go embeds. I have a series of patches ready to upstream to MMv1 that will allow this, but the google3 paths must be fixed before otherwise we'd break the library on subsequent imports.

BEGIN_PUBLIC
END_PUBLIC

PiperOrigin-RevId: 837417036