Retire `resource.CustomTemplate` in favor of `customTemplate $`.
Calling a templating method on resources means that the resource object currently means that Resources need to be aware of template locations, filesystem considerations, and in the future `fs.FS` objects.

On the other hand if we move this to the `google` package, those "global" template functions will have a way to pass through `fs.FS` objects.

Before removing the `CustomTemplate` method from Resource and Type, I've also double-checked that our private overrides do not use these.